Revert "route naming and avgRating type change"
This reverts commit 3ff3ede71075aa1ea7e6816d82a33d9130ae4e15.

+++ src/main/java/com/ssst/homefinderbackend/data/entity/RealEstateEntity.java
@@ -44,7 +44,7 @@ public class RealEstateEntity {
     private String address;
 
     @Column(name = "avg_rating", nullable = false)
-    private Double avgRating;
+    private Integer avgRating;
 
     @CreationTimestamp
     @Column(name = "created_at", updatable = false)
